requirements.txt
openai>=1.30.0
python-dotenv>=1.0.1
requests>=2.32.0
SpeechRecognition>=3.10.4
PyAudio>=0.2.14
gTTS>=2.5.1
playsound==1.2.2
pyttsx3>=2.90
openai-whisper>=20231117
sounddevice>=0.4.6
numpy>=1.26.0
google-api-python-client>=2.130.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
